refactor: add custom NumericFormat component
