+++ app/components/averageInPrice/AverageDown.tsx
@@ -1,10 +1,9 @@
+import StockProvider from "@/app/context/stock/StockProvider";
 import { Stock } from "@/app/types/stockTypes";
-import AddPurchase from "./AddPurchaseView";
+import AddPurchaseContainer from "./AddPurchaseContainer";
 import AdditionalStocks from "./AdditionalStocks";
 import HoldingStock from "./HoldingStock";
-import StockProvider from "@/app/context/stock/StockProvider";
 import ResultContainer from "./ResultContainer";
-import AddPurchaseContainer from "./AddPurchaseContainer";
 
 const AverageDown = ({ stockList }: { stockList: Stock[] }) => {
   return (
