+++ app/components/customUI/SearchBox.tsx
